Make Particle common parent of Lepton and Meson
We want to simplify our `Lepton` and `Meson` classes and therefore
share a few common class members. Additionally we can use that
opportunity to include `pdgid` as a new identifier that makes it
easier to distinguish particles in the future. This also solves
the problem of how to enter lepton mixing generations in the command
line.

So in the future this should not be

  --generation \\mu,e,\\tau

any longer but instead

  --generation 13,11,15

which makes it much easier to distinguish.

Additionally we compare now particles based on `pdgid` and not on
mass. This has two advantages:

  - for one we can now distinguish charges (anti-particles)
  - we avoid indetermined behaviour with different precisions of
    masses because of precision or similar