Uncomment U2-search to speed up testing
The useful feature to search for given ctau and mass for a fitting
U2 is now by default uncommented.

It is part of the last *Example* comments in the main routine and
can be thought of as motiviation for future tests. Right now the
behaviour is simply not desired as I am just comparing calculations
with others.

Later @TODO would be to make this search configureable via a command
line option.