Use PDG ID as command line argument instead of LaTeX string
Using a LaTeX string like `\\mu` is unusual and prone to errors. A
better solution is to use the PDG ID as became standard for decades
and this commit fixes that.

The command line argument is still parsed as a TString as multiple
PDG IDs can be supplied using `,` and no whitespace.