Store decay channels as a vector of pdg IDs instead of names
When partial widths are calculated the result is stored in the HNL
object for later reference. Originally this was saved as a TString
but a more flexible way is to simply denote the PDG ID directly.

This of course needs to be parsed when outputting but that is more
modular and flexible than parsing TStrings.