Allocate particle construction to ParticleCatalogue
The new ParticleCatalogue class takes care of constructing all
leptons and mesons. It is therefore easy to use to get a certain PDG
ID within a function and unifies the construction of particles.

Currently it is restricted to positive PDG IDs.

Additionally in this commit we do not give custom names to particles
but strictly assign them the LaTeX label of the pdgIdToLaTeX function.

There is a remaining @TODO: Currently the auxfunctions.* contains
this function and it is identically existing as a private member in
the Particle class in order to avoid a dependence on auxfunctions.*

In the long term we should probably have a designated PDG ID class that
handles this.