Correct neutral vector meson partialwidth
An additional |V_ud|^2 factor was removed.

To make it more readable the g_h^2 and x_h^2 and k_h^2 values were
now already pre-calculated as well. That means: stored already in
a ghsq, xhsq and khsq field instead of calculating it on the fly
and storing it in a `temp` construct.

This also allowed the saving of one temporary element called `temp2`
which gives just a tiny edge on memory efficiency.