Add contribution section to README
The decision has been made to use a standard (could be any really)
for the C++ style. The choice was `LLVM` which is documented at

  http://llvm.org/docs/CodingStandards.html

and suggestion was also included to use a pre-commit file. However
for that I inserted now `pre-commit-clang-format` to the `.gitignore`
file in order make it easier for everyone to use it and _not_ commit
it.