Make searching for U2 configurable
An optional command line argument to supply a ctau value for which
a U2 should be found can now be stated. Alternatatively a list of
masses and ctau's can be supplied and parsed in the same argument.

The file needs to be formated like:

 float float

where the first float corresponds to the mass and the second to the
ctau.